pyproject.toml
[tool.poetry]
name = "dz-14"
version = "0.1.0"
description = ""
authors = ["9-Bogdan <127949078+9-Bogdan@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "dz_14"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
alembic = "^1.12.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.22"
pydantic = "^2.4.2"
fastapi-mail = "^1.4.1"
fastapi-limiter = "^0.1.5"
cloudinary = "^1.36.0"
pydantic-settings = "^2.0.3"
libgravatar = "^1.0.4"
python-dotenv = "^1.0.0"
pytest = "^7.4.3"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
